// src/components/Model/Rider.ts
export interface rider {
    id: number;
    riderName: string;
    riderAddress1: string;
    riderAddress2: string;
    riderAddress3: string;
    riderAddress4: string;
    riderCity: string;
    riderState: string;
    riderPostalCode: string;
    drivingLicenseFile: string;
    aadharPANFile: string;
    bankPANFile:string;
    l1Approve:string
    isActive:string;
    email: string;
    bloodGroup:string;
    phone:string;
    rating:String;
    photo:String;
  }